Add gnome-initial-setup user to systemd-journal group
When eos-installer hits an error, it automatically runs eos-diagnostics
to capture more details, and gives the user a link to view the output
and/or saves it to the USB stick.  When run from a standalone
eosinstaller image, or when the user chooses Reformat during the FBE on
a live image, it runs as the gnome-initial-setup user, which was
previously neither a member of systemd-journal nor of adm, so could not
read the journal.

Since gnome-initial-setup is root-equivalent anyway – it can create
user accounts in the administrator group – allowing it read-only access
to the journal seems okay.

https://phabricator.endlessm.com/T19950